+++ src/main/java/com/example/movie15/domain/movie/controller/MovieController.java
@@ -1,8 +1,10 @@
 package com.example.movie15.domain.movie.controller;
 
+import com.example.movie15.domain.movie.dto.MovieReviewsResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -67,6 +69,30 @@ public class MovieController {
 	public ResponseEntity<String> deleteMovie(@PathVariable Long movieId) {
 		movieService.deleteMovie(movieId);
 		return ResponseEntity.status(HttpStatus.OK).body("삭제완료");
+	}
+
+	/**
+	 * 영화 ID에 해당하는 리뷰 목록을 조회.
+	 *
+	 * @param movieId 조회할 영화의 ID
+	 * @return 영화에 대한 리뷰 목록을 포함하는 ResponseEntity. 영화에 대한 리뷰가 없으면 빈 리스트를 반환.
+	 */
+	@GetMapping("/{movieId}/reviews")
+	public ResponseEntity<Page<MovieReviewsResponseDto>> findMovieReviews(
+			@PathVariable Long movieId,
+			@RequestParam(defaultValue = "0") int page,
+			@RequestParam(defaultValue = "10") int size,
+			@RequestParam(defaultValue = "createdAt") String sortBy,
+			@RequestParam(defaultValue = "desc") String direction
+	) {
+		Sort sort = direction.equalsIgnoreCase("desc") ?
+				Sort.by(sortBy).descending() : Sort.by(sortBy).ascending();
+
+		Pageable pageable = PageRequest.of(page, size, sort);
+
+		Page<MovieReviewsResponseDto> reviewList = movieService.findMovieReviews(movieId, pageable);
+
+		return ResponseEntity.status(HttpStatus.OK).body(reviewList);
 	}
 
 }
