+++ src/main/java/com/example/movie15/domain/movie/repository/MovieRepository.java
@@ -1,10 +1,5 @@
 package com.example.movie15.domain.movie.repository;
 
-
-import java.util.List;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
